+++ app/src/main/java/com/andrerinas/headunitrevived/utils/Settings.kt
@@ -253,9 +253,29 @@ class Settings(context: Context) {
         get() = prefs.getBoolean("auto-start-self-mode", false)
         set(value) { prefs.edit().putBoolean("auto-start-self-mode", value).apply() }
 
+    var autoStartBluetoothDeviceName: String
+        get() = prefs.getString("auto-start-bt-name", "")!!
+        set(value) { prefs.edit().putString("auto-start-bt-name", value).apply() }
+
+    var autoStartBluetoothDeviceMac: String
+        get() = prefs.getString("auto-start-bt-mac", "")!!
+        set(value) { prefs.edit().putString("auto-start-bt-mac", value).apply() }
+
     var appLanguage: String
         get() = prefs.getString("app-language", "")!!
         set(value) { prefs.edit().putString("app-language", value).apply() }
+
+    var mediaVolumeOffset: Int
+        get() = prefs.getInt("media-volume-offset", 0)
+        set(value) { prefs.edit().putInt("media-volume-offset", value).apply() }
+
+    var assistantVolumeOffset: Int
+        get() = prefs.getInt("assistant-volume-offset", 0)
+        set(value) { prefs.edit().putInt("assistant-volume-offset", value).apply() }
+
+    var navigationVolumeOffset: Int
+        get() = prefs.getInt("navigation-volume-offset", 0)
+        set(value) { prefs.edit().putInt("navigation-volume-offset", value).apply() }
 
     @SuppressLint("ApplySharedPref")
     fun commit() {